+++ scm-plugins/scm-hg-plugin/src/main/java/sonia/scm/repository/spi/HgModifyCommand.java
@@ -41,11 +41,13 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
+import java.util.regex.Pattern;
 
 @SuppressWarnings("java:S3252") // it is ok for javahg classes to access static method of subtype
 public class HgModifyCommand implements ModifyCommand {
 
   private static final Logger LOG = LoggerFactory.getLogger(HgModifyCommand.class);
+  static final Pattern HG_MESSAGE_PATTERN = Pattern.compile(".*\\[SCM\\](?: Error:)? (.*)");
 
   private final HgCommandContext context;
   private final HgWorkingCopyFactory workingCopyFactory;
@@ -128,8 +130,12 @@ public class HgModifyCommand implements ModifyCommand {
       com.aragost.javahg.commands.PullCommand pullCommand = PullCommand.on(workingCopy.getCentralRepository());
       workingCopyFactory.configure(pullCommand);
       return pullCommand.execute(workingCopy.getDirectory().getAbsolutePath());
-    } catch (Exception e) {
-      throw new IntegrateChangesFromWorkdirException(context.getScmRepository(),
+    } catch (ExecutionException e) {
+      throw IntegrateChangesFromWorkdirException
+        .withPattern(HG_MESSAGE_PATTERN)
+        .forMessage(context.getScmRepository(), e.getMessage());
+    } catch (IOException e) {
+      throw new InternalRepositoryException(context.getScmRepository(),
         String.format("Could not pull modify changes from working copy to central repository for branch %s", request.getBranch()),
         e);
     }