+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/CreateGroupDto.java
@@ -22,43 +22,28 @@
  * SOFTWARE.
  */
 
-package sonia.scm;
+package sonia.scm.api.v2.resources;
 
-import java.util.Collections;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import sonia.scm.util.ValidationUtil;
+
+import javax.validation.constraints.Pattern;
 import java.util.List;
 
-import static java.util.stream.Collectors.joining;
+/**
+ * This class is currently only used in the openapi scheme
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+public class CreateGroupDto {
 
-public class NotFoundException extends ExceptionWithContext {
-
-  private static final long serialVersionUID = 1710455380886499111L;
-
-  private static final String CODE = "AGR7UzkhA1";
-
-  public NotFoundException(Class<?> type, String id) {
-    this(Collections.singletonList(new ContextEntry(type, id)));
-  }
-
-  public NotFoundException(String type, String id) {
-    this(Collections.singletonList(new ContextEntry(type, id)));
-  }
-
-  public static NotFoundException notFound(ContextEntry.ContextBuilder contextBuilder) {
-    return new NotFoundException(contextBuilder.build());
-  }
-
-  private NotFoundException(List<ContextEntry> context) {
-    super(context, createMessage(context));
-  }
-
-  @Override
-  public String getCode() {
-    return CODE;
-  }
-
-  private static String createMessage(List<ContextEntry> context) {
-    return context.stream()
-      .map(c -> c.getType().toLowerCase() + " with id " + c.getId())
-      .collect(joining(" in ", "could not find ", ""));
-  }
+  @Pattern(regexp = ValidationUtil.REGEX_NAME)
+  private String name;
+  private String description;
+  private String type;
+  private List<String> members;
+  private boolean external;
 }