+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/ResourceLinks.java
@@ -24,14 +24,14 @@
 
 package sonia.scm.api.v2.resources;
 
-import sonia.scm.repository.NamespaceAndName;
 import sonia.scm.security.gpg.UserPublicKeyResource;
 
 import javax.inject.Inject;
 import java.net.URI;
 import java.net.URISyntaxException;
 
-@SuppressWarnings("squid:S1192") // string literals should not be duplicated
+@SuppressWarnings("squid:S1192")
+  // string literals should not be duplicated
 class ResourceLinks {
 
   private final ScmPathInfoStore scmPathInfoStore;
@@ -273,13 +273,13 @@ class ResourceLinks {
   }
 
   AutoCompleteLinks autoComplete() {
-    return new AutoCompleteLinks (scmPathInfoStore.get());
+    return new AutoCompleteLinks(scmPathInfoStore.get());
   }
 
-  static class AutoCompleteLinks  {
+  static class AutoCompleteLinks {
     private final LinkBuilder linkBuilder;
 
-    AutoCompleteLinks (ScmPathInfo pathInfo) {
+    AutoCompleteLinks(ScmPathInfo pathInfo) {
       linkBuilder = new LinkBuilder(pathInfo, AutoCompleteResource.class);
     }
 
@@ -485,16 +485,20 @@ class ResourceLinks {
       branchLinkBuilder = new LinkBuilder(pathInfo, RepositoryRootResource.class, RepositoryResource.class, BranchRootResource.class);
     }
 
-    String self(NamespaceAndName namespaceAndName, String branch) {
-      return branchLinkBuilder.method("getRepositoryResource").parameters(namespaceAndName.getNamespace(), namespaceAndName.getName()).method("branches").parameters().method("get").parameters(branch).href();
+    String self(String namespace, String name, String branch) {
+      return branchLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("branches").parameters().method("get").parameters(branch).href();
     }
 
-    public String history(NamespaceAndName namespaceAndName, String branch) {
-      return branchLinkBuilder.method("getRepositoryResource").parameters(namespaceAndName.getNamespace(), namespaceAndName.getName()).method("branches").parameters().method("history").parameters(branch).href();
+    public String history(String namespace, String name, String branch) {
+      return branchLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("branches").parameters().method("history").parameters(branch).href();
     }
 
     public String create(String namespace, String name) {
       return branchLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("branches").parameters().method("create").parameters().href();
+    }
+
+    public String delete(String namespace, String name, String branch) {
+      return branchLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("branches").parameters().method("delete").parameters(branch).href();
     }
   }
 
@@ -510,11 +514,11 @@ class ResourceLinks {
     }
 
     public String changesets(String namespace, String name) {
-      return toTemplateParams(incomingLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("incoming").parameters().method("incomingChangesets").parameters("source","target").href());
+      return toTemplateParams(incomingLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("incoming").parameters().method("incomingChangesets").parameters("source", "target").href());
     }
 
     public String changesets(String namespace, String name, String source, String target) {
-      return incomingLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("incoming").parameters().method("incomingChangesets").parameters(source,target).href();
+      return incomingLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("incoming").parameters().method("incomingChangesets").parameters(source, target).href();
     }
 
     public String diff(String namespace, String name) {
@@ -591,6 +595,7 @@ class ResourceLinks {
     ModificationsLinks(ScmPathInfo pathInfo) {
       modificationsLinkBuilder = new LinkBuilder(pathInfo, RepositoryRootResource.class, RepositoryResource.class, ModificationsRootResource.class);
     }
+
     String self(String namespace, String name, String revision) {
       return modificationsLinkBuilder.method("getRepositoryResource").parameters(namespace, name).method("modifications").parameters().method("get").parameters(revision).href();
     }