+++ CHANGELOG.md
@@ -6,11 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+- Delete branches directly in the UI ([#1422](https://github.com/scm-manager/scm-manager/pull/1422))
+- Lookup command which provides further repository information ([#1415](https://github.com/scm-manager/scm-manager/pull/1415))
+- Include messages from scm protocol in modification or merge errors ([#1420](https://github.com/scm-manager/scm-manager/pull/1420))
+- Enhance trace api to accepted status codes ([#1430](https://github.com/scm-manager/scm-manager/pull/1430))
+
 ### Changed
 - Send mercurial hook callbacks over separate tcp socket instead of http ([#1416](https://github.com/scm-manager/scm-manager/pull/1416))
 
 ### Fixed
+- Missing close of hg diff command ([#1417](https://github.com/scm-manager/scm-manager/pull/1417))
 - Error on repository initialization with least-privilege user ([#1414](https://github.com/scm-manager/scm-manager/pull/1414))
+- Adhere to git quiet flag ([#1421](https://github.com/scm-manager/scm-manager/pull/1421))
+- Resolve svn binary diffs properly [#1427](https://github.com/scm-manager/scm-manager/pull/1427)
+
+## [2.9.1] - 2020-11-11
+### Fixed
+- German translation for repositories view
+
 
 ## [2.9.0] - 2020-11-06
 ### Added